Fixed compiler warnings about implied type casts.
git-svn-id: https://svn.linuxsampler.org/svn/linuxsampler/trunk@3706 d2c4447f-3961-4fff-bec3-96d84be9abe5

+++ src/engines/AbstractEngineChannel.cpp
@@ -3,9 +3,9 @@
  *   LinuxSampler - modular, streaming capable sampler                     *
  *                                                                         *
  *   Copyright (C) 2003,2004 by Benno Senoner and Christian Schoenebeck    *
- *   Copyright (C) 2005-2008 Christian Schoenebeck                         *
- *   Copyright (C) 2009-2012 Christian Schoenebeck and Grigor Iliev        *
- *   Copyright (C) 2012-2017 Christian Schoenebeck and Andreas Persson     *
+ *   Copyright (C) 2005-2020 Christian Schoenebeck                         *
+ *   Copyright (C) 2009-2012 Grigor Iliev                                  *
+ *   Copyright (C) 2012-2017 Andreas Persson                               *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -276,7 +276,7 @@ namespace LinuxSampler {
 
     uint AbstractEngineChannel::GetMidiInputPortCount() {
         Sync< ArrayList<MidiInputPort*> > connections = midiInputs.back();
-        return connections->size();
+        return (uint) connections->size();
     }
 
     MidiInputPort* AbstractEngineChannel::GetMidiInputPort(uint index) {
